+++ CHANGELOG.md
@@ -35,7 +35,7 @@ Version template:
 ### Fixed
 
 * Sharded solr monitoring throws exception in Alfresco 6.2
-* Fixed and refactored code for solr, so that it works for ASS>=2.0.0. Improved handling of solrconfig.xml for solr6.
+* Fixed and refactored code for solr, so that it works for ASS>=2.0.0. Improved handling of solrconfig.xml for solr6 [#74]
 
 [#74] https://github.com/xenit-eu/alfred-telemetry/pull/74